+++ examples/fuzz-ssl-server.py
@@ -1,18 +1,18 @@
 #!/usr/bin/env python2
 # designed for use with boofuzz
 #
-# minimal example which fuzzes a remote TLS server
+# minimal example which fuzzes a remote HTTPS server
+# use this as a starting point for SSL/TLS server fuzzing
 
-import ssl
 from boofuzz import *
 
 # If you don't want to verify remote certificate, create a SSLContext.
 # WARNING: You will be vulnerable to a man-in-the-middle attack!
+#   import ssl
 #   ctx = ssl.create_default_context(ssl.Purpose.SERVER_AUTH)
 #   ctx.check_hostname = False
 #   ctx.verify_mode = ssl.CERT_NONE
 
-boofuzz/socket_connection.py
 session = Session(
     target=Target(
         connection=SocketConnection(
@@ -21,7 +21,7 @@ session = Session(
             proto='ssl',
             server_hostname="google.de",
             # sslcontext=ctx,
-            ),
+        ),
     ),
 )
 s_initialize("GET Request")